split up CLI
